# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: NEXUS Super Social Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /ai/feed:
    get:
      operationId: getAiFeed
      tags: [ai]
      summary: AI-personalized content feed
      parameters:
        - in: query
          name: userId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiFeed"
...
  /ai/suggestions:
    get:
      operationId: getAiSuggestions
      tags: [ai]
      summary: AI user/content suggestions
      parameters:
        - in: query
          name: userId
          schema: { type: integer }
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AiSuggestions"
  /ai/moderation:
    post:
      operationId: moderateContent
      tags: [ai]
      summary: AI content moderation check
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ModerationInput"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ModerationResult"

  # ── ADMIN ─────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    User:
      type: object
      required: [id, username, displayName, email, createdAt]
      properties:
        id: { type: integer }
        username: { type: string }
        displayName: { type: string }
        email: { type: string }
        bio: { type: ["string", "null"] }
        avatarUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        followersCount: { type: integer }
        followingCount: { type: integer }
        postsCount: { type: integer }
        isVerified: { type: boolean }
        isFollowing: { type: boolean }
        status: { type: string, enum: [active, suspended] }
        createdAt: { type: string }
...
    Post:
      type: object
      required: [id, author, content, type, likesCount, commentsCount, createdAt]
      properties:
        id: { type: integer }
        author:
          $ref: "#/components/schemas/User"
        content: { type: string }
        type: { type: string, enum: [photo, video, text] }
        mediaUrl: { type: ["string", "null"] }
        thumbnailUrl: { type: ["string", "null"] }
        tags: { type: array, items: { type: string } }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        sharesCount: { type: integer }
        isLiked: { type: boolean }
        isFlagged: { type: boolean }
        midnightOnly: { type: boolean }
        createdAt: { type: string }
...
    Reel:
      type: object
      required: [id, author, videoUrl, caption, likesCount, viewsCount, createdAt]
      properties:
        id: { type: integer }
        author:
          $ref: "#/components/schemas/User"
        videoUrl: { type: string }
        thumbnailUrl: { type: ["string", "null"] }
        caption: { type: string }
        audioTrack: { type: ["string", "null"] }
        duration: { type: integer }
        likesCount: { type: integer }
        commentsCount: { type: integer }
        viewsCount: { type: integer }
        isLiked: { type: boolean }
        tags: { type: array, items: { type: string } }
        createdAt: { type: string }
        views24h: { type: integer }
...
    Group:
      type: object
      required: [id, name, description, membersCount, isPrivate, createdAt]
      properties:
        id: { type: integer }
        name: { type: string }
        description: { type: string }
        avatarUrl: { type: ["string", "null"] }
        coverUrl: { type: ["string", "null"] }
        membersCount: { type: integer }
        postsCount: { type: integer }
        isPrivate: { type: boolean }
        isMember: { type: boolean }
        category: { type: string }
        createdAt: { type: string }
        privacyLevel: { type: string }
        joinType: { type: string }
        groupType: { type: string }
        icon: { type: string }
        themeColor: { type: string }
        maxMembers: { type: integer }
        settings: { type: object }
        creatorId: { type: ["integer", "null"] }
...
    AiFeed:
      type: object
      required: [posts, reels, suggestedUsers]
      properties:
        posts:
          type: array
          items:
            $ref: "#/components/schemas/Post"
        reels:
          type: array
          items:
            $ref: "#/components/schemas/Reel"
        suggestedUsers:
          type: array
          items:
            $ref: "#/components/schemas/User"
        echoScore:
          type: integer
          description: "0-100 share of the feed's personalization signal concentrated in a single tag (echo chamber detector)"
        echoTopTag:
          type: string
          nullable: true
...
    AiSuggestions:
      type: object
      required: [users, groups, topics]
      properties:
        users:
          type: array
          items:
            $ref: "#/components/schemas/User"
        groups:
          type: array
          items:
            $ref: "#/components/schemas/Group"
        topics:
          type: array
          items: { type: string }
    ModerationInput:
      type: object
      required: [content]
      properties:
        content: { type: string }
        contentType: { type: string, enum: [text, image_url] }
    ModerationResult:
      type: object
      required: [safe, confidence, flags]
      properties:
        safe: { type: boolean }
        confidence: { type: number }
        flags:
          type: array
          items: { type: string }
        recommendation: { type: string }

    # ── Admin ─────────────────────────────────────────────────────────────────
...